Fix trie_test include and fix memory bug with 32-bit int on 64-bit platforms using clang
Because void* is 8 bytes on 64-bit platforms and 32-bit in is only 4, the four bytes preceding the int foo would get overwritten with \0s. In my case the 4 bytes preceding foo were the four low bytes of *trie, which would corrupt the trie pointer

+++ trie_test.c
@@ -1,4 +1,4 @@
-#include "trie.c"
+#include "trie.h"
 #include <stdio.h>
 
 int main(void)
@@ -20,40 +20,40 @@ int main(void)
    if (!trie_insert(trie, "fof", (void*)4))
       printf("failed to insert fof\n");
 
-   int foo, fooo, fof, bar, bab, empty, null;
+   long foo, fooo, fof, bar, bab, empty, null;
 
    if (trie_fetch(trie, "foo", (void*)&foo))
-      printf("foo = %d\n", foo);
+      printf("foo = %ld\n", foo);
    else
       printf("foo not found\n");
 
    if (trie_fetch(trie, "fooo", (void*)&fooo))
-      printf("fooo = %d\n", fooo);
+      printf("fooo = %ld\n", fooo);
    else
       printf("fooo not found\n");
 
    if (trie_fetch(trie, "bar", (void*)&bar))
-      printf("bar = %d\n", bar);
+      printf("bar = %ld\n", bar);
    else
       printf("bar not found\n");
 
    if (trie_fetch(trie, "fof", (void*)&fof))
-      printf("fof = %d\n", fof);
+      printf("fof = %ld\n", fof);
    else
       printf("fof not found\n");
 
    if (trie_fetch(trie, "bab", (void*)&bab))
-      printf("bab = %d\n", bab);
+      printf("bab = %ld\n", bab);
    else
       printf("bab not found\n");
 
    if (trie_fetch(trie, "", (void*)&empty))
-      printf("empty = %d\n", empty);
+      printf("empty = %ld\n", empty);
    else
       printf("empty not found\n");
 
    if (trie_fetch(trie, NULL, (void*)&null))
-      printf("NULL key = %d\n", null);
+      printf("NULL key = %ld\n", null);
    else
       printf("NULL key not found\n");
